requirements.txt
# 치과용 얼굴 좌표 인식 프로그램 필수 패키지

# 컴퓨터 비전
opencv-python>=4.8.0
mediapipe>=0.10.0

# 수치 연산 및 데이터 처리
numpy>=1.24.0

# 시각화
matplotlib>=3.7.0

# 유틸리티
Pillow>=10.0.0
